refactor : NextWalEntity 영속성 객체 <-> NextWal 도메인 객체 분리로 인한 변경
