+++ src/main/java/backend/wal/wal/nextwal/application/port/in/GetNextWalUseCase.java
@@ -1,7 +1,7 @@
 package backend.wal.wal.nextwal.application.port.in;
 
 import backend.wal.wal.nextwal.domain.NextWals;
-import backend.wal.wal.nextwal.domain.aggregate.NextWal;
+import backend.wal.wal.nextwal.domain.NextWal;
 
 public interface GetNextWalUseCase {
