+++ src/main/java/backend/wal/wal/nextwal/application/port/in/NextWalSettingUseCase.java
@@ -2,7 +2,7 @@ package backend.wal.wal.nextwal.application.port.in;
 
 import backend.wal.wal.common.domain.WalCategoryType;
 import backend.wal.wal.nextwal.domain.NextWals;
-import backend.wal.wal.nextwal.domain.aggregate.NextWal;
+import backend.wal.wal.nextwal.domain.NextWal;
 
 import java.util.Set;
