+++ src/main/java/backend/wal/wal/todaywal/adapter/NextWalSettingAdapter.java
@@ -4,7 +4,7 @@ import backend.wal.wal.common.domain.WalCategoryType;
 import backend.wal.wal.nextwal.application.port.in.GetNextWalUseCase;
 import backend.wal.wal.nextwal.application.port.in.NextWalSettingUseCase;
 import backend.wal.wal.nextwal.domain.NextWals;
-import backend.wal.wal.nextwal.domain.aggregate.NextWal;
+import backend.wal.wal.nextwal.domain.NextWal;
 import backend.wal.wal.todaywal.application.port.out.NextWalSettingPort;
 
 import org.springframework.stereotype.Component;
